requirements.txt
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
flask-pydantic-spec==0.3.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
nested-lookup==0.2.25
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.1
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
typing_extensions==4.4.0
Werkzeug==2.2.2
